C64.emu updates
* C64.emu: Update core to VICE 3.6.1
* C64.emu: Remove Virtual Device Traps option and manage it automatically when setting TDE
* C64.emu: Fix aligned pointer types in VICII rendering code
* C64.emu: Use fixed size array to store CPU trap functions
* C64.emu: Allow selecting files from any path in media control menu

+++ C64.emu/src/vice/core/flash040core.c
@@ -184,7 +184,7 @@ inline static int flash_program_byte(flash040_context_t *flash040_context, unsig
 inline static int flash_write_operation_status(flash040_context_t *flash040_context)
 {
     return ((flash040_context->program_byte ^ 0x80) & 0x80)   /* DQ7 = inverse of programmed data */
-           | ((maincpu_clk & 2) << 5)                         /* DQ6 = toggle bit (2 us) */
+           | ((int)(maincpu_clk & 2) << 5)                    /* DQ6 = toggle bit (2 us) */
            | (1 << 5)                                         /* DQ5 = timeout */
     ;
 }
