+++ C64.emu/src/vice/diskimage.h
@@ -27,6 +27,7 @@
 #ifndef VICE_DISKIMAGE_H
 #define VICE_DISKIMAGE_H
 
+#include <sys/types.h>
 #include "types.h"
 
 #define D64_FILE_SIZE_35   174848        /* D64 image, 35 tracks */
@@ -158,6 +159,6 @@ extern void disk_image_attach_log(const disk_image_t *image, signed int lognum,
                                   unsigned int unit, unsigned int drive);
 extern void disk_image_detach_log(const disk_image_t *image, signed int lognum,
                                   unsigned int unit, unsigned int drive);
-extern uint32_t disk_image_size(const disk_image_t *image);
+extern off_t disk_image_size(const disk_image_t *image);
 
 #endif
