+++ C64.emu/src/vice/diskimage/fsimage.h
@@ -59,6 +59,6 @@ extern int fsimage_read_sector(const struct disk_image_s *image, uint8_t *buf,
                                const struct disk_addr_s *dadr);
 extern int fsimage_write_sector(struct disk_image_s *image, const uint8_t *buf,
                                 const struct disk_addr_s *dadr);
-extern uint32_t fsimage_size(const disk_image_t *image);
+extern off_t fsimage_size(const disk_image_t *image);
 
 #endif
